requirements.txt
streamlit
openai
faiss-cpu
python-docx
pdfplumber
tqdm
python-dotenv
